// src/platform/packages/shared/kbn-discover-utils/src/types.d.ts
type DiscoverSearchHit = SearchHit<Record<string, unknown>>;
export interface EsHitRecord extends Omit<DiscoverSearchHit, '_index' | '_id' | '_source'> {
    _index?: DiscoverSearchHit['_index'];
    _id?: DiscoverSearchHit['_id'];
    _source?: DiscoverSearchHit['_source'];
}
/**
 * This is the record/row of data provided to our Data Table
 */
export interface DataTableRecord {
    /**
     * A unique id generated by index, id and routing of a record
     */
    id: string;
    /**
     * The document returned by Elasticsearch for search queries
     */
    raw: EsHitRecord;
    /**
     * A flattened version of the ES doc or data provided by SQL, aggregations ...
     */
    flattened: Record<string, unknown>;
    /**
     * Determines that the given doc is the anchor doc when rendering view surrounding docs
     */
    isAnchor?: boolean;
}